Slightly Deconstructed and stripped code
Using some code from my old plugin I have chopped out everything that wasn't necessary and have kept only the necessary for reminding myself what we will need to include in the future and what we have to simplify.

This is going to be actual agony but its worth it.